Create a new context and query for the search_form
