+++ lib/siwapp/query.ex
@@ -25,11 +25,4 @@ defmodule Siwapp.Query do
     query
     |> where([q], ilike(field(q, ^string_field), ^search))
   end
-
-  def name_email_or_id(query, value) do
-    query
-    |> where([q], ilike(q.name, ^"%#{value}%"))
-    |> or_where([q], ilike(q.email, ^"%#{value}%"))
-    |> or_where([q], ilike(q.identification, ^"%#{value}%"))
-  end
 end
