+++ lib/siwapp/search/search_query.ex
@@ -1,0 +1,13 @@
+defmodule Siwapp.Search.SearchQuery do
+  @moduledoc """
+  Search Queries
+  """
+  import Ecto.Query
+
+  def name_email_or_id(query, value) do
+    query
+    |> where([q], ilike(q.name, ^"%#{value}%"))
+    |> or_where([q], ilike(q.email, ^"%#{value}%"))
+    |> or_where([q], ilike(q.identification, ^"%#{value}%"))
+  end
+end
