+++ lib/siwapp_web/live/invoices_live/index.ex
@@ -1,7 +1,8 @@
 defmodule SiwappWeb.InvoicesLive.Index do
   @moduledoc false
   use SiwappWeb, :live_view
-  alias Siwapp.Invoices
+  alias Siwapp.Invoices.Invoice
+  alias Siwapp.{Invoices, Search}
   alias SiwappWeb.PageView
 
   def mount(_params, _session, %{id: "home"} = socket) do
@@ -67,7 +68,7 @@ defmodule SiwappWeb.InvoicesLive.Index do
   end
 
   def handle_event("search", params, socket) do
-    invoices = Invoices.filtered(params["search_input"])
+    invoices = Search.filters(Invoice, params["search_input"])
     {:noreply, assign(socket, :invoices, invoices)}
   end
